Split up giant AstDsl.java file
